Extract migration error and result types to their own files
