+++ WordPress/src/main/java/org/wordpress/android/localcontentmigration/LocalContentEntity.kt
@@ -5,9 +5,6 @@ import org.wordpress.android.fluxc.model.PostModel
 import org.wordpress.android.fluxc.model.QuickStartStatusModel
 import org.wordpress.android.fluxc.model.QuickStartTaskModel
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.localcontentmigration.LocalContentEntityData.Companion.IneligibleReason
-import org.wordpress.android.localcontentmigration.LocalMigrationResult.Failure
-import org.wordpress.android.localcontentmigration.LocalMigrationResult.Success
 import org.wordpress.android.models.ReaderPostList
 
 enum class LocalContentEntity(private val isIdentifiable: Boolean = false) {
@@ -50,49 +47,4 @@ sealed class LocalContentEntityData {
             ;
         }
     }
-}
-
-sealed class LocalMigrationError {
-    sealed class ProviderError: LocalMigrationError() {
-        data class NullValueFromQuery(val forEntity: LocalContentEntity): ProviderError()
-        data class NullCursor(val forEntity: LocalContentEntity): ProviderError()
-        data class ParsingException(val forEntity: LocalContentEntity): ProviderError()
-    }
-    data class Ineligibility(val reason: IneligibleReason): LocalMigrationError()
-    sealed class FeatureDisabled: LocalMigrationError() {
-        object SharedLoginDisabled: FeatureDisabled()
-        object UserFlagsDisabled: FeatureDisabled()
-    }
-    sealed class MigrationAlreadyAttempted: LocalMigrationError() {
-        object SharedLoginAlreadyAttempted: MigrationAlreadyAttempted()
-        object UserFlagsAlreadyAttempted: MigrationAlreadyAttempted()
-    }
-    sealed class PersistenceError: LocalMigrationError() {
-        object FailedToSaveSites: PersistenceError()
-        object FailedToSaveUserFlags: PersistenceError()
-    }
-    object NoUserFlagsFoundError: LocalMigrationError()
-}
-
-sealed class LocalMigrationResult<out T: LocalContentEntityData, out E: LocalMigrationError> {
-    data class Success<T: LocalContentEntityData>(val value: T): LocalMigrationResult<T, Nothing>()
-    data class Failure<E: LocalMigrationError>(val error: E): LocalMigrationResult<Nothing, E>()
-}
-
-fun <T: LocalContentEntityData, U: LocalContentEntityData, E: LocalMigrationError> LocalMigrationResult<T, E>
-        .thenWith(next: (T) -> LocalMigrationResult<U, E>) = when (this) {
-    is Success -> next(this.value)
-    is Failure -> this
-}
-
-fun <T: LocalContentEntityData> LocalMigrationResult<LocalContentEntityData, LocalMigrationError>
-        .then(next: () -> LocalMigrationResult<T, LocalMigrationError>) = when (this) {
-    is Success -> next()
-    is Failure -> this
-}
-
-fun <T: LocalContentEntityData, E: LocalMigrationError> LocalMigrationResult<T, E>
-        .otherwise(handleError: (E) -> Unit) = when (this) {
-    is Success -> Unit
-    is Failure -> handleError(this.error)
 }
